+++ lib/screens/invoicing/bank_details_screen.dart
@@ -128,6 +128,7 @@ class _BankDetailsScreenState extends State<BankDetailsScreen> {
                     const SizedBox(height: 24),
                     TextFormField(
                       controller: _bankNameController,
+                      keyboardAppearance: Brightness.light,
                       decoration: const InputDecoration(
                         labelText: 'Bank Name',
                         prefixIcon: Icon(AppIcons.bank),
@@ -137,6 +138,7 @@ class _BankDetailsScreenState extends State<BankDetailsScreen> {
                     const SizedBox(height: 16),
                     TextFormField(
                       controller: _accountNameController,
+                      keyboardAppearance: Brightness.light,
                       decoration: const InputDecoration(
                         labelText: 'Account Name',
                         prefixIcon: Icon(AppIcons.user),
@@ -146,6 +148,7 @@ class _BankDetailsScreenState extends State<BankDetailsScreen> {
                     const SizedBox(height: 16),
                     TextFormField(
                       controller: _sortCodeController,
+                      keyboardAppearance: Brightness.light,
                       decoration: const InputDecoration(
                         labelText: 'Sort Code',
                         hintText: 'XX-XX-XX',
@@ -157,6 +160,7 @@ class _BankDetailsScreenState extends State<BankDetailsScreen> {
                     const SizedBox(height: 16),
                     TextFormField(
                       controller: _accountNumberController,
+                      keyboardAppearance: Brightness.light,
                       decoration: const InputDecoration(
                         labelText: 'Account Number',
                         prefixIcon: Icon(AppIcons.tag),
@@ -167,6 +171,7 @@ class _BankDetailsScreenState extends State<BankDetailsScreen> {
                     const SizedBox(height: 16),
                     TextFormField(
                       controller: _paymentTermsController,
+                      keyboardAppearance: Brightness.light,
                       decoration: const InputDecoration(
                         labelText: 'Payment Terms',
                         prefixIcon: Icon(AppIcons.clock),