+++ lib/screens/tools/battery_load_test_screen.dart
@@ -181,6 +181,7 @@ class _BatteryLoadTestScreenState extends State<BatteryLoadTestScreen> {
             // Battery Capacity
             TextFormField(
               controller: _batteryCapacityController,
+              keyboardAppearance: Brightness.light,
               decoration: InputDecoration(
                 labelText: 'Installed Battery Capacity (Ah)',
                 prefixIcon: Icon(AppIcons.batteryFull),
@@ -213,6 +214,7 @@ class _BatteryLoadTestScreenState extends State<BatteryLoadTestScreen> {
             // Standby Current
             TextFormField(
               controller: _standbyCurrentController,
+              keyboardAppearance: Brightness.light,
               decoration: InputDecoration(
                 labelText: 'Standby Current I\u2081 (A)',
                 prefixIcon: Icon(AppIcons.flash),
@@ -237,6 +239,7 @@ class _BatteryLoadTestScreenState extends State<BatteryLoadTestScreen> {
             // Alarm Current
             TextFormField(
               controller: _alarmCurrentController,
+              keyboardAppearance: Brightness.light,
               decoration: InputDecoration(
                 labelText: 'Alarm Current I\u2082 (A)',
                 prefixIcon: Icon(AppIcons.notification),